+++ Daedalus/Rotation/PrometheusCore/Abilities/PrometheusAbilities.cs
@@ -87,4 +87,7 @@ public static class PrometheusAbilities
 
     // --- Role ---
     public static readonly AbilityBehavior HeadGraze = new() { Action = RoleActions.HeadGraze, Toggle = cfg => cfg.RangedShared.EnableHeadGraze };
+
+    public static readonly AbilityBehavior SecondWind = new() { Action = RoleActions.SecondWind, Toggle = cfg => cfg.RangedShared.EnableSecondWind };
+    public static readonly AbilityBehavior Tactician = new() { Action = MCHActions.Tactician, Toggle = cfg => cfg.RangedShared.EnablePartyMitigation };
 }